requirements.txt
fastapi
uvicorn[standard]
streamlit
sentence-transformers
faiss-cpu
python-multipart
beautifulsoup4
lxml
selenium
jinja2
pydantic
aiofiles
requests
numpy
tqdm
transformers
accelerate
sentencepiece
